Implement bounded-memory storage paths and borrowed-buffer columnar writes
- Streaming RG-by-RG rewrite for delta compaction (compact_streaming_v4),
  compressed-row-group deletes (rewrite_v4_active_rows), and DROP COLUMN
  (rewrite_v4_drop_columns): peak memory is O(largest Row Group + delta)
  instead of the whole table
- Footer-only ADD COLUMN on mmap-only V4 tables with all-NULL synthesis on
  read; schema changes compact any pending delta first
- Chunk appends by row_group_size so single large batches never create
  oversized Row Groups
- Upgrade in-memory String/Binary/StringDict offsets to u64 (disk format
  keeps u32 per Row Group, existing files stay compatible)
- store_columnar builds columns directly from borrowed &str/&[u8] buffers via
  StorageEngine::write_typed_columns, eliminating per-element String/Vec
  allocation and the typed intermediate copy; FTS string retention only when
  enabled
- Fix _id leakage and missing-column padding in mmap read paths exposed by
  footer-only ADD COLUMN
- Add Rust and Python regression tests; refresh storage architecture docs

+++ apexbase/src/storage/bloom.rs
@@ -142,7 +142,7 @@ impl ColumnBloomIndex {
     /// Build bloom index from string column data
     pub fn build_from_strings(
         column_name: &str,
-        offsets: &[u32],
+        offsets: &[u64],
         data: &[u8],
         row_group_size: usize,
         false_positive_rate: f64,
@@ -402,7 +402,7 @@ mod tests {
 
     #[test]
     fn test_column_bloom_index() {
-        let offsets = vec![0u32, 5, 11, 17, 23, 29]; // 5 strings
+        let offsets = vec![0u64, 5, 11, 17, 23, 29]; // 5 strings
         let data = b"appleorangebananagrapemeloncoding";
 
         let index = ColumnBloomIndex::build_from_strings(
